fix(featurecounts): Run against the annotation name
This allows us to just run all of the crosses without worrying about
trying to magically filter them. They can be quickly processed after
with a find, without trying to wrangle the meta data etc.

I also found the we could add the tuple(meta) to each of the grohmm runs
and maybe go down that path, but I don't see any reason to keep that
grouping around this far down stream.

This needs to be solidified and upstreamed to modules